+++ src/main/java/net/frozenblock/wilderwild/tag/WWBiomeTags.java
@@ -317,6 +317,7 @@ public final class WWBiomeTags {
 	public static final TagKey<Biome> HAS_NETHER_LAVA_GEYSER = bind("feature/has_nether_lava_geyser");
 	public static final TagKey<Biome> HAS_HYDROTHERMAL_VENT = bind("feature/has_hydrothermal_vent");
 	public static final TagKey<Biome> HAS_HYDROTHERMAL_VENT_RARE = bind("feature/has_hydrothermal_vent_rare");
+	public static final TagKey<Biome> HAS_SULFUR_SPRING = bind("feature/has_sulfur_spring");
 	public static final TagKey<Biome> HAS_TAIGA_FOREST_ROCK = bind("feature/has_taiga_forest_rock");
 	public static final TagKey<Biome> HAS_SNOW_PILE = bind("feature/has_snow_pile");
 	public static final TagKey<Biome> HAS_MOSS_PATH = bind("feature/has_moss_path");